backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
alpaca-py
